+++ apps/api/src/services/data-provider/eod-historical-data/eod-historical-data.service.ts
@@ -163,10 +163,10 @@ export class EodHistoricalDataService implements DataProviderInterface {
       ).json<any>();
 
       return response.reduce(
-        (result, { close, date }) => {
-          if (isNumber(close)) {
+        (result, { adjusted_close, date }) => {
+          if (isNumber(adjusted_close)) {
             result[this.convertFromEodSymbol(symbol)][date] = {
-              marketPrice: close
+              marketPrice: adjusted_close
             };
           } else {
             Logger.error(